requirements.txt
python-dotenv
openai
rich
halo
huggingface_hub
